+++ InsCorp/web/src/main/java/kabilova/tu/inscorp/web/admin/VehicleTypeUpDel.java
@@ -1,6 +1,8 @@
 package kabilova.tu.inscorp.web.admin;
 
+import kabilova.tu.inscorp.model.vehicle.Vehicle;
 import kabilova.tu.inscorp.model.vehicle.VehicleType;
+import kabilova.tu.inscorp.server.web.VehicleServer;
 import kabilova.tu.inscorp.server.web.VehicleTypeServer;
 
 import javax.servlet.RequestDispatcher;
@@ -24,25 +26,48 @@ public class VehicleTypeUpDel extends HttpServlet {
         request.setCharacterEncoding("UTF-8");
 
         Integer id = Integer.parseInt(request.getParameter("vehicleType"));
-        VehicleType vehicleType = new kabilova.tu.inscorp.model.vehicle.VehicleType();
-        vehicleType.setId(id);
 
-        VehicleTypeServer vehicleTypeServer = new VehicleTypeServer(vehicleType);
-
-        if(request.getServletPath().equals("/vehicleTypeUpdate")){
-            String newVehicleType = request.getParameter("newVehicleType");
-            vehicleType.setVehicleType(newVehicleType);
-
-            vehicleTypeServer.update();
-
-            request.setAttribute("errmsg", "Успешен запис!");
+        if(id==0) {
+            request.setAttribute("errmsg", "Моля, попълнете всички полета!");
             RequestDispatcher view = request.getRequestDispatcher("admin/AdminPanelMsg.jsp");
-            view.forward(request,response);
-        } else if(request.getServletPath().equals("/vehicleTypeDelete")){
-            vehicleTypeServer.delete();
-            request.setAttribute("errmsg", "Успешен запис!");
-            RequestDispatcher view = request.getRequestDispatcher("admin/AdminPanelMsg.jsp");
-            view.forward(request,response);
+            view.forward(request, response);
+        } else {
+            VehicleType vehicleType = new VehicleType();
+            vehicleType.setId(id);
+
+            VehicleTypeServer vehicleTypeServer = new VehicleTypeServer(vehicleType);
+            VehicleServer vehicleServer = new VehicleServer(new Vehicle());
+
+            if (request.getServletPath().equals("/vehicleTypeUpdate")) {
+                String newVehicleType = request.getParameter("newVehicleType");
+                vehicleType.setVehicleType(newVehicleType);
+
+                vehicleTypeServer.update();
+
+                request.setAttribute("errmsg", "Успешен запис!");
+                RequestDispatcher view = request.getRequestDispatcher("admin/AdminPanelMsg.jsp");
+                view.forward(request, response);
+            } else if (request.getServletPath().equals("/vehicleTypeDelete")) {
+                boolean canDelete = true;
+
+                for (Vehicle v : vehicleServer.read()) {
+                    if (v.getVehicleType().getId() == id) {
+                        canDelete = false;
+                    }
+                }
+
+                if (canDelete) {
+                    vehicleTypeServer.delete();
+
+                    request.setAttribute("errmsg", "Успешен запис!");
+                    RequestDispatcher view = request.getRequestDispatcher("admin/AdminPanelMsg.jsp");
+                    view.forward(request, response);
+                } else {
+                    request.setAttribute("errmsg", "Невъзможно изтриване на тип МПС, по които има въведени МПС!");
+                    RequestDispatcher view = request.getRequestDispatcher("admin/AdminPanelMsg.jsp");
+                    view.forward(request, response);
+                }
+            }
         }
     }
 }
